+++ service/connectparticipant/doc.go
@@ -4,13 +4,9 @@
 // requests to Amazon Connect Participant Service.
 //
 // Amazon Connect is a cloud-based contact center solution that makes it easy
-// to set up and manage a customer contact center and provide reliable customer
-// engagement at any scale.
-//
-// Amazon Connect enables customer contacts through voice or chat.
-//
-// The APIs described here are used by chat participants, such as agents and
-// customers.
+// to set up and manage a customer contact center. Amazon Connect enables customer
+// contacts through voice or chat. Use the Amazon Connect Participant Service
+// to manage chat participants, such as agents and customers.
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/connectparticipant-2018-09-07 for more information on this service.
 //