+++ service/detective/api.go
@@ -73,6 +73,10 @@ func (c *Detective) AcceptInvitationRequest(input *AcceptInvitationInput) (req *
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - ConflictException
 //     The request attempted an invalid action.
 //
@@ -161,6 +165,10 @@ func (c *Detective) BatchGetGraphMemberDatasourcesRequest(input *BatchGetGraphMe
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -245,6 +253,10 @@ func (c *Detective) BatchGetMembershipDatasourcesRequest(input *BatchGetMembersh
 // API operation BatchGetMembershipDatasources for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -350,6 +362,10 @@ func (c *Detective) CreateGraphRequest(input *CreateGraphInput) (req *request.Re
 // API operation CreateGraph for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - ConflictException
 //     The request attempted an invalid action.
@@ -477,6 +493,10 @@ func (c *Detective) CreateMembersRequest(input *CreateMembersInput) (req *reques
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -580,6 +600,10 @@ func (c *Detective) DeleteGraphRequest(input *DeleteGraphInput) (req *request.Re
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -680,6 +704,10 @@ func (c *Detective) DeleteMembersRequest(input *DeleteMembersInput) (req *reques
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - ConflictException
 //     The request attempted an invalid action.
 //
@@ -771,6 +799,10 @@ func (c *Detective) DescribeOrganizationConfigurationRequest(input *DescribeOrga
 // API operation DescribeOrganizationConfiguration for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -870,6 +902,10 @@ func (c *Detective) DisableOrganizationAdminAccountRequest(input *DisableOrganiz
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -962,6 +998,10 @@ func (c *Detective) DisassociateMembershipRequest(input *DisassociateMembershipI
 // API operation DisassociateMembership for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - ConflictException
 //     The request attempted an invalid action.
@@ -1068,6 +1108,10 @@ func (c *Detective) EnableOrganizationAdminAccountRequest(input *EnableOrganizat
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -1154,6 +1198,10 @@ func (c *Detective) GetMembersRequest(input *GetMembersInput) (req *request.Requ
 // API operation GetMembers for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -1245,6 +1293,10 @@ func (c *Detective) ListDatasourcePackagesRequest(input *ListDatasourcePackagesI
 // API operation ListDatasourcePackages for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -1391,6 +1443,10 @@ func (c *Detective) ListGraphsRequest(input *ListGraphsInput) (req *request.Requ
 // API operation ListGraphs for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -1539,6 +1595,10 @@ func (c *Detective) ListInvitationsRequest(input *ListInvitationsInput) (req *re
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -1685,6 +1745,10 @@ func (c *Detective) ListMembersRequest(input *ListMembersInput) (req *request.Re
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -1828,6 +1892,10 @@ func (c *Detective) ListOrganizationAdminAccountsRequest(input *ListOrganization
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -1965,6 +2033,10 @@ func (c *Detective) ListTagsForResourceRequest(input *ListTagsForResourceInput) 
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -2056,6 +2128,10 @@ func (c *Detective) RejectInvitationRequest(input *RejectInvitationInput) (req *
 // API operation RejectInvitation for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - ConflictException
 //     The request attempted an invalid action.
@@ -2152,6 +2228,10 @@ func (c *Detective) StartMonitoringMemberRequest(input *StartMonitoringMemberInp
 // API operation StartMonitoringMember for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - ConflictException
 //     The request attempted an invalid action.
@@ -2255,6 +2335,10 @@ func (c *Detective) TagResourceRequest(input *TagResourceInput) (req *request.Re
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -2341,6 +2425,10 @@ func (c *Detective) UntagResourceRequest(input *UntagResourceInput) (req *reques
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -2426,6 +2514,10 @@ func (c *Detective) UpdateDatasourcePackagesRequest(input *UpdateDatasourcePacka
 // API operation UpdateDatasourcePackages for usage and error information.
 //
 // Returned Error Types:
+//
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
 //
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
@@ -2528,6 +2620,10 @@ func (c *Detective) UpdateOrganizationConfigurationRequest(input *UpdateOrganiza
 //
 // Returned Error Types:
 //
+//   - AccessDeniedException
+//     The request issuer does not have permission to access this resource or perform
+//     this operation.
+//
 //   - InternalServerException
 //     The request was valid but failed because of a problem with the service.
 //
@@ -2629,6 +2725,83 @@ func (s AcceptInvitationOutput) String() string {
 // value will be replaced with "sensitive".
 func (s AcceptInvitationOutput) GoString() string {
 	return s.String()
+}
+
+// The request issuer does not have permission to access this resource or perform
+// this operation.
+type AccessDeniedException struct {
+	_            struct{}                  `type:"structure"`
+	RespMetadata protocol.ResponseMetadata `json:"-" xml:"-"`
+
+	// The SDK default error code associated with the access denied exception.
+	ErrorCode *string `type:"string" enum:"ErrorCode"`
+
+	// The SDK default explanation of why access was denied.
+	ErrorCodeReason *string `type:"string"`
+
+	Message_ *string `locationName:"Message" type:"string"`
+
+	// The error code associated with the access denied exception.
+	SubErrorCode *string `type:"string" enum:"ErrorCode"`
+
+	// An explanation of why access was denied.
+	SubErrorCodeReason *string `type:"string"`
+}
+
+// String returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s AccessDeniedException) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation.
+//
+// API parameter values that are decorated as "sensitive" in the API will not
+// be included in the string output. The member name will be present, but the
+// value will be replaced with "sensitive".
+func (s AccessDeniedException) GoString() string {
+	return s.String()
+}
+
+func newErrorAccessDeniedException(v protocol.ResponseMetadata) error {
+	return &AccessDeniedException{
+		RespMetadata: v,
+	}
+}
+
+// Code returns the exception type name.
+func (s *AccessDeniedException) Code() string {
+	return "AccessDeniedException"
+}
+
+// Message returns the exception's message.
+func (s *AccessDeniedException) Message() string {
+	if s.Message_ != nil {
+		return *s.Message_
+	}
+	return ""
+}
+
+// OrigErr always returns nil, satisfies awserr.Error interface.
+func (s *AccessDeniedException) OrigErr() error {
+	return nil
+}
+
+func (s *AccessDeniedException) Error() string {
+	return fmt.Sprintf("%s: %s\n%s", s.Code(), s.Message(), s.String())
+}
+
+// Status code returns the HTTP status code for the request's response error.
+func (s *AccessDeniedException) StatusCode() int {
+	return s.RespMetadata.StatusCode
+}
+
+// RequestID returns the service's response RequestID for request.
+func (s *AccessDeniedException) RequestID() string {
+	return s.RespMetadata.RequestID
 }
 
 // An Amazon Web Services account that is the administrator account of or a